use_case_5/requirements.txt
openai
python-dotenv
python-docx
Pymupdf==1.25.5
pymupdf4llm==0.0.20


